% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{calc_cv_sgl}
\alias{calc_cv_sgl}
\title{Calculate crossvalidation to estimate lambda1.se & identify potential underpowered features}
\usage{
calc_cv_sgl(
  y,
  x,
  model = "sparse.grp.lasso",
  intercept = TRUE,
  seed_cv = 1234,
  folds_cv = 5
)
}
\value{
list containing lambda1.se and feature names excluding underpowered ones
}
\description{
Calculate crossvalidation to estimate lambda1.se & identify potential underpowered features
}
