% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_frobenius_norm}
\alias{calc_frobenius_norm}
\title{estimate frobenius norm of a matrix}
\usage{
calc_frobenius_norm(x)
}
\value{
mse double
}
\description{
estimate frobenius norm of a matrix
}
