% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{calc_lambda1.se}
\alias{calc_lambda1.se}
\title{Calculate lambda1.se model for crossvalidation}
\usage{
calc_lambda1.se(lambdas, error_cv)
}
\value{
lambda1.se
}
\description{
Calculate lambda1.se model for crossvalidation
}
