% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{calc_tau}
\alias{calc_tau}
\title{Calculate tau based on frobenius norm}
\usage{
calc_tau(x)
}
\value{
Frobenius norm of x
}
\description{
Calculate tau based on frobenius norm
}
\examples{
x <- matrix(1:20,5,4)
calc_tau(x)
}
