% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prior.R
\name{create_prior_network}
\alias{create_prior_network}
\title{creating a prior network on basis of the prior_map}
\usage{
create_prior_network(prior_map)
}
\arguments{
\item{prior_map}{data frame - edge list -wir column names  'A','B','layer_A','layer_B'}
}
\value{
igraph network
}
\description{
creating a prior network on basis of the prior_map
}
