% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{fetch_mappings}
\alias{fetch_mappings}
\title{extracting the relevant mapping information}
\usage{
fetch_mappings(node, mapping)
}
\arguments{
\item{node}{, String}

\item{mapping}{,  data.table with ids in thecolumns 1:2}
}
\value{
vector of feature names
}
\description{
extracting the relevant mapping information
}
