% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{fetch_var}
\alias{fetch_var}
\title{using the prior information to fetsh the right data for X}
\usage{
fetch_var(
  node,
  input_list,
  mapping_list,
  metainfo,
  main_layer = 1,
  sep = "___"
)
}
\arguments{
\item{node}{,}

\item{input_list}{,}

\item{mapping_list}{,}

\item{main_layer}{- default = 1}
}
\value{
X sample x feature matrix
}
\description{
using the prior information to fetsh the right data for X
}
