% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{infer_network}
\alias{infer_network}
\title{Infers a model for each node in the main layer}
\usage{
infer_network(
  input_list,
  mapping_list,
  metainfo,
  main_layer = 1,
  min_features = 2,
  sel_iterations = 0,
  core = 1
)
}
\arguments{
\item{input_list}{- list of omics data. First list element will be used as predictor}

\item{mapping_list}{- list of mappings between each data type one}

\item{metainfo}{- table of relation between mappings and input list}

\item{main_layer}{- which input data set represents the main layer (default = 1)}

\item{min_features}{- autoexclude models with less than 2 features (default = 2)}

\item{sel_iterations}{- run stability selection, defines number of iterations. if !=0 it will perform stability selection}

\item{core}{- to run network inference in parallel}
}
\value{
a network in form of an edge table
}
\description{
Infers a model for each node in the main layer
}
