% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{is_valid}
\alias{is_valid}
\title{check if data is valid}
\usage{
is_valid(x, min_features)
}
\arguments{
\item{x}{, matrix features in columns and samples in rows}

\item{min_features}{, default 5}
}
\value{
TRUE / FALSE
}
\description{
check if data is valid
}
