% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prior.R
\name{load_mapping}
\alias{load_mapping}
\title{loading mapping files}
\usage{
load_mapping(df, layers = NA)
}
\arguments{
\item{df}{data frame or mapping file}

\item{layers}{- default NA, if no colnames please specify your layer names here}
}
\value{
data frame containing all mappings
}
\description{
loading mapping files
}
