% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{parse_prior_groups}
\alias{parse_prior_groups}
\title{Extracts groups numbers}
\usage{
parse_prior_groups(names, sep = "\\\\___")
}
\value{
vector of numeric groups
}
\description{
Extracts groups numbers
}
\examples{
names <- c("methylation___cg0123123","rppa___MTOR")
infer_prior_groups(names)
}
