% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_kimono}
\alias{plot_kimono}
\title{generate igraph}
\usage{
plot_kimono(ig_kimono, title = "")
}
\arguments{
\item{ig_kimono}{igraph of kimono network}

\item{title}{plot title}
}
\value{
igraph
}
\description{
generate igraph
}
