% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{preprocess_data}
\alias{preprocess_data}
\title{remove na, scale}
\usage{
preprocess_data(y, x)
}
\arguments{
\item{y}{, vector of doubles}

\item{x}{, matrix features in columns and samples in rows}
}
\value{
x, y without na's
}
\description{
remove na, scale
}
