% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{stability_selection}
\alias{stability_selection}
\title{Stability selection and summary of multiple runs}
\usage{
stability_selection(y, x, nseeds)
}
\arguments{
\item{y}{data.table - feature to predict}

\item{nseeds}{- specifies how many iterations to run}

\item{X}{data.table - input features with prior names attached to features}
}
\value{
edge list for a given input y and x, and statistics on multiple runs
}
\description{
Stability selection and summary of multiple runs
}
