% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{to_igraph}
\alias{to_igraph}
\title{generate igraph}
\usage{
to_igraph(network, directed = TRUE)
}
\arguments{
\item{network}{edge list generated by kimono}

\item{directed}{if false the method will collaps all edges and remove free nodes as well as single directional edges.}
}
\value{
igraph
}
\description{
generate igraph
}
