% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sgl_model.R
\name{train_kimono_sgl}
\alias{train_kimono_sgl}
\title{Calculate crossvalidation to estimate lambda1.se & identify potential underpowered features}
\usage{
train_kimono_sgl(y, x, model = "sparse.grp.lasso", intercept = TRUE, ...)
}
\value{
edge list for a given input y and x
}
\description{
Calculate crossvalidation to estimate lambda1.se & identify potential underpowered features
}
