% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimono.R
\name{kimono}
\alias{kimono}
\title{Run Kimono - Knowledge-guIded Multi-Omic Netowrk inference}
\usage{
kimono(
  input_data,
  prior_network,
  min_features = 2,
  sel_iterations = 0,
  core = 1,
  specific_layer = NULL,
  DEBUG = FALSE,
  scdata = FALSE,
  ...
)
}
\arguments{
\item{prior_network}{- igraph with prior information}

\item{min_features}{- autoexclude models with less than 2 features (defaul}

\item{core}{- if core != 1 kimono will perform an parallell computation}

\item{specific_layer}{- specify layer}

\item{DEBUG}{- only run on subset}

\item{scdata}{- if single cell data}

\item{input_list}{- list of omics data. First list element will be used as predictor}
}
\value{
a network in form of an edge table
}
\description{
Run Kimono - Knowledge-guIded Multi-Omic Netowrk inference
}
